% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_early_time.R
\name{early.time}
\alias{early.time}
\title{Early time for a deterministic projects}
\usage{
early.time(precedence, duration)
}
\arguments{
\item{precedence}{a matrix that indicates the order of precedence between activities.  If the value \eqn{(i,j)} is equal to 1 then \eqn{i} precedes \eqn{j}.}

\item{duration}{vector with the duración for each activities.}
}
\value{
Early time vector.
}
\description{
This function calculates the early time for one project.
}
\examples{
precedence<-matrix(c(0,0,1,0,0,0,0,0,1,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0),nrow=5,ncol=5,byrow=TRUE)
duration<-c(3,2,1,1.5,4.2)
early.time(precedence,duration)
}
\references{
\describe{
  \item{}{Burke, R. (2013). Project management: planning and control techniques. New Jersey, USA.}
}
}
