% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_organize.R
\name{organize}
\alias{organize}
\title{Organize project activities}
\usage{
organize(precedence)
}
\arguments{
\item{precedence}{A matrix that indicates the order of precedence between activities.  If the value \eqn{(i,j)} is equal to 1 then \eqn{i} precedes \eqn{j}.}
}
\value{
A list containing:
\itemize{
\item{Precedence: }{ ordered precedence matrix.}
\item{Order: }{ new activities values.}
}
}
\description{
This function organizes the activities of a project, in such a way that if i precedes j then i is less strict than j.
}
\examples{
precedence<-matrix(c(0,0,1,0,0,0,0,0,1,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0),nrow=5,ncol=5,byrow=TRUE)
organize(precedence)
}
